What effects does huge Boris factor have?
Without the electron inertial effects, we need a Boris factor around 1e5 to see any parallel electric field (and even then it's small). If we do crank it up so high, what do we see? We'll drive a pair of runs overnight to see. Just 20s long, one with no Boris factor (E3=0) and one with it at 1e5.